Make sure `VersionRange::Empty#hash` works
The subclass does not respond to `#min` or `#max` defined, so if `#hash`
ends up being called, it results in a crash because the parent class
defines `#hash` using these attributes.

This can happen when checking whether an incompatibility has
been previously seen, if the incompatibility includes a
`VersionRange::Empty` term.

This commit fixes the issue by explicitly defining
`VersionRange::Empty#hash`.