Use better next dependency selection strategy
While working on migrating Bundler to use PubGrub, I'm noticing several
cases where, while PubGrub resolves correctly, its result is clearly
inferior to current Bundler's result and it's also slower.

Further investigation pointed to the decision making phase, where
Bundler is being smarter and choosing better dependencies that led into
finding a solution earlier.

Looking at PubGrub result, I noticed the problem was that the solution
includes ancient versions of gems nobdoy wants a fresh bundle to resolve
to.

For example, one realworld case was resolving to Rails 2.3.2 (in
exchange for choosing the recent 3.0.0 release of rack), while Bundler
properly sticks to modern Rails and to the previous 2.2.4 release of
Rack.

On another realworld case, PubGrub was taking about 20 seconds to find a
solution (reeeeeally slow for PubGrub standards) before finding a
solution involving... Rails 0.9.5, after 49 attempts.

My diagnosis of the issue was that the older a dependency is, the less
priority it should get in the decision making phase. And my criteria to
decide how old a dependency is is how many versions of the dependency
are higher than the dependency range upper bound.

With this criteria, these realworld cases resolve instantly and with the
same result as Molinillo.