Expose `SolveFailure` incompatibility
In Bundler, we normally try resolution with as many locked gems as
possible from the Gemfile.lock file, but if resolution fails in that
"conservative mode", we automatically unlock any locked gems that cause
conflicts and retry in a more "free mode". We do this by inspecting the
initial `SolveFailure` and figuring out if any locked gems caused the
conflicts.

Exposising SolveFailure's incompatibility is useful for this.