Use `to_s` min and max representation when displaying ranges
In Bundler, we consider "1.12.5" and "1.12.5-x86_64-linux" different for
resolution purposes, but equal when representing versions in error
messages and so on.

Using strict min and max equality for range string representation leads
to same weird ranges reading like "< 1.12.5, > 1.12.5".

It seems best to use the equality of string representations of mix and
max to decide whether to display "=" or "<, >" for representing ranges.