Use less formal wording for forbidden terms
"Forbidden" reads as overly technical and may be confusing since it's
more often used as in the context of authorization. Basically it had "an
illegal operation was performed" vibes.

This changes the wording from "x is forbidden" to "x cannot be used",
and also drops "every version of" when appropriate.

Wording suggested by André Arko

Co-authored-by: Daniel Colson <composerinteralia@github.com>