Detect infinite loops when resolving conflicts
We've been getting several reports about Bundler hanging infinitely when
resolving.

I believe this is a bug in pub_grub, but while I isolate a test case
and fix, I think it's an improvement already to detect infinite loops
when resolving conflicts and bail out.

We can remove this once the culprit is properly addressed.