Optimize `#next_package_to_try` by not sorting
A large app I work on was taking around a minute to `bundle update`, so
I profiled it and found that `#sort_versions_by_preferred` was a large
hotspot (~30% of total runtime).

I initially assumed that these calls must be inside
`#choose_package_version`, where the sorting of `#versions_for` is
important because a specific package version is picked. However, the 30%
of runtime inside `#versions_for` was actually only the calls made
inside `#next_package_to_try` (where the sorting is not used!).

This commit extracts `#sorted_versions_for`, which only fetches the
pre-sorted versions that match the range but does not perform additional
sorting afterwards. This prevents the costly sorts that are inevitably
discarded.

Also note that Bundler has similarly extracted the same method, but
named `select_sorted_versions`.